Improved available roles when remove them

+++ modules/Settings/Roles/actions/Delete.php
@@ -19,10 +19,9 @@ class Settings_Roles_Delete_Action extends Settings_Vtiger_Basic_Action
 		$moduleModel = Settings_Vtiger_Module_Model::getInstance($qualifiedModuleName);
 		$recordModel = Settings_Roles_Record_Model::getInstanceById($recordId);
 		$transferToRole = Settings_Roles_Record_Model::getInstanceById($transferRecordId);
-		if ($recordModel && $transferToRole) {
+		if ($recordModel && $transferToRole && strpos($transferToRole->get('parentrole'), $recordModel->get('parentrole')) === false) {
 			$recordModel->delete($transferToRole);
 		}
-
 		$redirectUrl = $moduleModel->getDefaultUrl();
 		header("location: $redirectUrl");
 	}
